requirements.txt
# deepFluxUniHelp - Dependencies
# Phase 1 - Base setup

# Backend API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# RAG & LLM
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
langchain-chroma>=0.1.2
sentence-transformers>=2.2.0

# Document processing
pypdf>=3.0.0
python-docx>=1.0.0
docx2txt>=0.8

# Frontend (Streamlit)
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
reportlab>=4.0.0
httpx>=0.25.0
pytest>=7.0.0
